Change values from int to long for hate
Implement some dragon valley raids ai
Thanks to l2j

Implement vangara script zone for future usage

Core:

compile and copy paste core.jar

Data:

Delete scripts folder and paste from this rev

+++ L2J_SunriseProject_Core/java/l2r/gameserver/model/actor/instance/L2FortCommanderInstance.java
@@ -70,7 +70,7 @@ public class L2FortCommanderInstance extends L2DefenderInstance
 	}
 	
 	@Override
-	public void addDamageHate(L2Character attacker, int damage, int aggro)
+	public void addDamageHate(L2Character attacker, int damage, long aggro)
 	{
 		if (attacker == null)
 		{
